Add support to control RTL-SDR bias tee.
The RTL-SDR V3 dongle allows software control of the bias tee DC injection.
This new option allows control of that feature.

Because not every distribution of librtlsdr supports bias tee setting,
it is made a build option (RTLSDR_BIASTEE)  and left disabled by default.
For Debian, the minimum librtlsdr version is specified.